Move `TestConfig` to `tests` module (#798)
* moving test-config to test-module

* fixing the test
